Move DataBoundConfiguratior and permit BaseConfigurator.TypePair as a part of Beta API
